+++ tests/test_exceptions.py
@@ -1,6 +1,9 @@
 """Tests for exceptions and requests"""
+import osgeo  # https://gis.stackexchange.com/questions/417733/unable-to-import-python-rasterio-package-even-though-it-is-installed
 import pytest
 from shapely import Polygon
+
+osgeo.__version__  # So nox doesn't remove it
 
 import pynldas2 as nldas
 from pynldas2 import InputRangeError, InputTypeError, InputValueError